Improve error when runner passed missing file path
When the runner is passed a non existing file, the runner attempts to
evaluate the path as a javascript expression, which can lead to a non
obvious error message.

Explicitly add that an expression was not found as a file to the error
message thrown when an expression passed to the runner is evaluated.